feat: (password) pattern-attribute
With 'type="password"' you can use the 'pattern' attribute to define a regular expression that the password must match to be considered valid.

Add a 'pattern' attribute to the password 'input' element to require the input match: '[a-z0-5]{8,}'

The above is a regular expression which matches eight or more lowercase letters or the digits '0' to '5'. Then, remove the 'minlength' attribute, and try it out.