feat: input-max min age
Add another 'label' after the first, with the text 'Input your age (years):' . Then, nest an 'input' with the 'type' of 'number'.

As we do not want users under the age of 13 to register, add a 'min' attribute to the 'input' with a value of '13'. Also, we can probably assume users over the age of 120 will not register; add a 'max' attribute with a value of '120'.

Now, if someone tries to submit the form with values outside of the range, a warning will appear, and the form will not submit. Give it a try.