feat: formatting borders
To make the 'input' and 'textarea' elements blend in with the 'background' theme, set their 'background-color' to '#0a0a23'. Then, give them a '1px', 'solid' border with a color of '#0a0a23'.
_______________________
Rather than creating separate style rules for the border-width, border-color, and border-style properties, it is simpler to use the CSS shorthand technique that specifies a value for each of these 3 properties to a border property as a space-separated list.
This uniformly styles each side of the context box with a border of the specific width, color, and style.